Migrated from nuspec to sdk style nuget package
This is a preparation for adding dotnet tool support for this
nuget package. All metadata is moved from the Gettext.Tools.nuspec
to the Directory.Build.props file which gets applied to all projects
on a lower directory level.

A new project src/Gettext.Tools was introduced which contains the
package id and additional files that are only needed for the nuget
PackageType "reference". Additionally, the project is used to
build the nuget package with the "dotnet pack" command which is
now used by the Makefile.

The resulting package is exactly the same as if it was created with
the nuspec file.