Added dotnet tool support
With these changes, the Gettext.Tools can additionally
be installed as a dotnet tool. This makes it possible to
install it as global tool and directly use it from the
command line.

A dotnet tool can register only a single command. Therefore,
the gettext tool itself must be passed as a parameter, including
individual options, to the "top-level" Gettext.Tools command. e.g.
Gettext.Tools -- <TOOL_NAME> [tool specific options]

It was needed to create a separate nuget package for the tools
functionality as the PackAsTool property removes all other package
types from the generated nuspec file.